🐛 Fix circular import bug, add try-except block to install chrmore to export images using kaleido, and update sactter plot notebook
