+++ .gitignore
@@ -165,3 +165,4 @@ cython_debug/
 # Tests
 test_results/
 docs/api_examples/outputs/
+docs/api_examples/iframe_figures/
