+++ pyproject.toml
@@ -38,6 +38,7 @@ dependencies = [
     "acore",
     "dash-cytoscape",
     "pydantic",
+    "nbformat>=4.2.0",
 ]
 
 [project.optional-dependencies]
